+++ core/src/test/scala/org/apache/spark/WordSpecSparkTestSuite.scala
@@ -1,0 +1,47 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.spark
+
+import org.scalatest.wordspec.AnyWordSpec
+
+class WordSpecSparkTestSuite
+  extends AnyWordSpec
+    with SparkTestSuite
+    with LocalSparkContext {
+
+  override def beforeAll(): Unit = {
+    super.beforeAll()
+    sc = new SparkContext("local", "test")
+  }
+
+  override def afterAll(): Unit = {
+    if (sc != null) {
+      sc.stop()
+      sc = null
+    }
+    super.afterAll()
+  }
+
+  "A RDD" when {
+    "empty" should {
+      "have size 0" in {
+        assert(sc.emptyRDD[Int].count() === 0)
+      }
+    }
+  }
+}
